[#14519] - Added tests
